+++ src/fcntl.rs
@@ -1,5 +1,12 @@
 use crate::errno::Errno;
 use libc::{self, c_int, c_uint, size_t, ssize_t};
+#[cfg(any(
+    target_os = "netbsd",
+    target_os = "macos",
+    target_os = "ios",
+    target_os = "dragonfly",
+))]
+use std::ffi::CStr;
 use std::ffi::OsString;
 #[cfg(not(target_os = "redox"))]
 use std::os::raw;
@@ -561,13 +568,11 @@ pub fn fcntl(fd: RawFd, arg: FcntlArg) -> Result<c_int> {
             #[cfg(any(target_os = "dragonfly", target_os = "netbsd", target_os = "macos", target_os = "ios"))]
             F_GETPATH(path) => {
                 let mut buffer = vec![0; libc::PATH_MAX as usize];
-                let res = libc::fcntl(fd, libc::F_GETPATH, buffer.as_ptr());
+                let res = libc::fcntl(fd, libc::F_GETPATH, buffer.as_mut_ptr());
                 let ok_res = Errno::result(res)?;
-                let len = buffer.iter().position(|b| *b == 0).unwrap();
-                buffer.truncate(len as usize);
-                buffer.shrink_to_fit();
-                *path = PathBuf::from(OsString::from_vec(buffer));
-                ok_res
+                let optr = CStr::from_bytes_until_nul(&buffer).unwrap();
+                *path = PathBuf::from(OsString::from(optr.to_str().unwrap()));
+                return Ok(ok_res)
             },
         }
     };